Remove unnecessary routes
These routes were, more or less, required, when the app was going to all
be written in PHP, or when I had quite a different idea of what it would
be. Now, as it's a much smaller application, one composed of one part in
PHP and the other in Vue.js, the routes are no longer required.