Update the GIMP charity background image
This change increases the font size of the charity's name so that it is
easier to read. This change was required owing to the background image
no longer being rendered as a background image, rather using an img tag.
Given that, the full image is displayed and the original text was much
smaller.